TST Fix error with numpy 1.10
Numpy 1.10 raises an exception when substracting two arrays with
different dtypes. Error was: Cannot cast ufunc subtract output from
dtype('float64') to dtype('int64') with casting rule 'same_kind'.

+++ nilearn/mass_univariate/tests/test_permuted_least_squares.py
@@ -137,7 +137,7 @@ def test_permuted_ols_check_h0_noeffect_labelswap(random_state=0):
     n_samples = 100
     # create dummy design with no effect
     target_var = rng.randn(n_samples, 1)
-    tested_var = np.arange(n_samples).reshape((-1, 1))
+    tested_var = np.arange(n_samples, dtype='f8').reshape((-1, 1))
     tested_var_not_centered = tested_var.copy()
     tested_var -= tested_var.mean(0)  # centered
     # permuted OLS
